+++ CHANGELOG.md
@@ -90,7 +90,15 @@ The present file will list all changes made to the project; according to the
 - `Reservation::displayReservationDay()`
 - `Reservation::displayReservationsForAnItem()`
 
-## [9.5.1] unreleased
+
+## [9.5.2] unreleased
+
+### API changes
+
+#### Removed
+
+- Ability to use SQL expressions as string in criterion values in SQL iterator (replaced by usage of `QueryExpression`).
+- Ability to delete a plugin image using `/front/pluginimage.send.php` script.
 
 ## [9.5.0] 2020-07-07